requirements.txt
torch
torchvision
timm
gradio
opencv-python
cryptography
pillow
numpy
